+++ src/Device/SerialPort.hpp
@@ -119,8 +119,8 @@ public:
 
   unsigned GetRxTimeout();
   virtual bool SetRxTimeout(unsigned Timeout);
-  virtual unsigned long GetBaudrate() const;
-  virtual unsigned long SetBaudrate(unsigned long BaudRate);
+  virtual unsigned GetBaudrate() const;
+  virtual unsigned SetBaudrate(unsigned BaudRate);
   virtual bool StopRxThread();
   virtual bool StartRxThread();
   void ProcessChar(char c);
